Add a lsp-warning for old tramp versions
In the author's experience, tramp 2.5.0-pre worked fine with old and new
implementation of lsp-tramp-connection. From the bug reports, reddit
messages and other information, it looked like emacs27.1 + tramp
>2.5.0-pre works fine

Added a Troubleshooting section to the remote doc with examples
of the errors. Suggested solution and tips to debug other
remote LSP errors.